requirements.txt
PyQt5>=5.15.0
Pillow>=10.0.0
pygame>=2.5.0
